requirements.txt
click>=8.0.0
GitPython>=3.1.0
colorama>=0.4.0